[SW-178][Prep for external backend] Remove duplicated code used by ProductRDDConverter and DataSet converter.
This commit also applies new convertor api to Dataset conversion and changes the DatasetConvertor to have the
same format as other convertors. This commit should not bring any semantic changes.

This commit also adds missing conversions from Dataset to H2OFrame key
