+++ core/src/main/scala/water/api/RDDs/RDDsHandler.scala
@@ -17,7 +17,7 @@
 package water.api.RDDs
 
 import org.apache.spark.SparkContext
-import org.apache.spark.h2o.converters.SupportedDataset.ProductFrameBuilder
+import org.apache.spark.h2o.converters.{DatasetConverter, H2OFrameFromRDDProductBuilder}
 import org.apache.spark.h2o.{H2OContext, H2OFrame}
 import org.apache.spark.mllib.regression.LabeledPoint
 import org.apache.spark.rdd.RDD
@@ -26,9 +26,9 @@ import water.api.Handler
 import water.exceptions.H2ONotFoundArgumentException
 
 /**
- * Handler for all RDD related queries
- */
-class RDDsHandler(val sc: SparkContext, val h2oContext: H2OContext) extends Handler {
+  * Handler for all RDD related queries
+  */
+class RDDsHandler(val sc: SparkContext, val h2oContext: H2OContext) extends Handler with DatasetConverter {
 
   def list(version: Int, s: RDDsV3): RDDsV3 = {
     val r = s.createAndFillImpl()
@@ -55,7 +55,7 @@ class RDDsHandler(val sc: SparkContext, val h2oContext: H2OContext) extends Hand
       // transform empty Seq in order to create empty H2OFrame
       h2oContext.asH2OFrame(sc.parallelize(Seq.empty[Int]),name)
     } else {
-       rdd.first() match {
+      rdd.first() match {
         case t if t.isInstanceOf[Double] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[Double]], name)
         case t if t.isInstanceOf[LabeledPoint] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[LabeledPoint]], name)
         case t if t.isInstanceOf[Boolean] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[Boolean]], name)
@@ -64,7 +64,7 @@ class RDDsHandler(val sc: SparkContext, val h2oContext: H2OContext) extends Hand
         case t if t.isInstanceOf[Float] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[Float]], name)
         case t if t.isInstanceOf[Long] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[Long]], name)
         case t if t.isInstanceOf[java.sql.Timestamp] => h2oContext.asH2OFrame(rdd.asInstanceOf[RDD[java.sql.Timestamp]], name)
-        case t if t.isInstanceOf[Product] => ProductFrameBuilder(sc, rdd.asInstanceOf[RDD[Product]], name).withDefaultFieldNames()
+        case t if t.isInstanceOf[Product] => H2OFrameFromRDDProductBuilder(h2oContext, rdd.asInstanceOf[RDD[Product]], name).withDefaultFieldNames()
         case t => throw new IllegalArgumentException(s"Do not understand type $t")
       }
     }
